requirements.txt
alembic==1.8.1
amqp==5.1.1
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
boto3==1.24.72
botocore==1.27.72
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
flasgger==0.9.5
Flask==2.2.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.16.0
kombu==5.2.4
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.17.1
marshmallow-sqlalchemy==0.28.1
mistune==2.0.4
numpy==1.23.3
packaging==21.3
Pillow==8.3.2
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.41
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
